Fixes to the camera, new textures, disabling scene editing
