+++ Aethra.RayTracer/Instructions/BoxWithSpheres.cs
@@ -67,6 +67,19 @@ namespace Aethra.RayTracer.Instructions
                 SpecularExponent = 50,
                 AmbientPower = 1
             };
+            
+            var abstractMaterial = new PbrMaterial(FloatColor.White,
+                Texture.LoadFrom(@"_Resources/Textures/Abstract_010_basecolor_green.jpg").ToInfo(0.25f),
+                Texture.LoadFrom(@"_Resources/Textures/Abstract_010_basecolor_green.jpg").ToInfo(0.25f),
+                Texture.LoadFrom(@"_Resources/Textures/Abstract_010_roughness.jpg").ToInfo(0.25f),
+                Texture.LoadFrom(@"_Resources/Textures/Abstract_010_normal.jpg").ToInfo(0.25f))
+            {
+                EmissionFactor = 1,
+                DiffuseCoefficient = 1,
+                Specular = 10,
+                SpecularExponent = 50,
+                AmbientPower = 1
+            };
 
             var reflectiveSphere = new Sphere(new Vector3(-0.5f, -1.5f, 3), 0.5f, reflectiveMaterial);
             var transparentSphere = new Sphere(new Vector3(0.5f, -1.5f, 3), 0.5f, transparentMaterial);
@@ -106,8 +119,8 @@ namespace Aethra.RayTracer.Instructions
             objects.Add(specialSphere3);
             objects.Add(new Plane(new Vector3(-2, 0, 0), new Vector3(1, 0, 0), reflectiveFloor));
             objects.Add(new Plane(new Vector3(2, 0, 0), new Vector3(-1, 0, 0), reflectiveFloor));
-            objects.Add(new Plane(new Vector3(5, -2f, 0), new Vector3(0, 1, 0), reflectiveFloor));
-            objects.Add(new Plane(new Vector3(5, 2f, 0), new Vector3(0, -1, 0), reflectiveFloor));
+            objects.Add(new Plane(new Vector3(0, -2f, 0), new Vector3(0, 1, 0), reflectiveFloor));
+            objects.Add(new Plane(new Vector3(0, 2f, 0), new Vector3(0, -1, 0), reflectiveFloor));
             objects.Add(new Plane(new Vector3(0, 2, 6), new Vector3(0, 0, -1), reflectiveFloor));
             objects.Add(new Plane(new Vector3(0, 2, -8), new Vector3(0, 0, 1), reflectiveFloor));
 
@@ -115,13 +128,11 @@ namespace Aethra.RayTracer.Instructions
 
             var samplerLights = new Sampler(new RegularGenerator(), new SquareDistributor(), 25, 1);
 
-            var camera = new PerspectiveCamera(renderTarget, new Vector3(0f, 0, -5), Vector3.Forward, Vector3.Up)
+            var camera = new PerspectiveCamera(renderTarget, new Vector3(0, 0, -3f), Vector3.Forward, Vector3.Up)
             {
-                //Sampler = sampler,
-                //SpecialColoring = (ray, hit) => FloatColor.FromNormal(hit.Normal)
+                Sampler = sampler,
                 MaxDepth = 5,
                 //ColorSpace = ColorSpace.Gamma
-                //SpecialColoring = (ray, hit) => FloatColor.FromVectorSafe(hit.Position)
             };
 
             Scene = new Scene(objects, camera,
@@ -130,32 +141,32 @@ namespace Aethra.RayTracer.Instructions
                     new PointLight
                     {
                         Position = new Vector3(-1.5f, -1.5f, 1),
-                        Color = FloatColor.Green,
-                         Sampler = samplerLights,
+                        Color = FloatColor.White,
+                        //Sampler = samplerLights,
                         Radius = 0,
                         Intensity = 0.5f
                     },
                     new PointLight
                     {
                         Position = new Vector3(1.5f, 1.5f, 1),
-                        Color = FloatColor.Red,
-                         Sampler = samplerLights,
+                        Color = FloatColor.White,
+                        //Sampler = samplerLights,
                         Radius = 0,
-                        Intensity = 0.6f
+                        Intensity = 0.5f
                     },
                     new PointLight
                     {
                         Position = new Vector3(1.5f, -1.5f, 1),
-                        Color = FloatColor.Blue,
-                         Sampler = samplerLights,
+                        Color = FloatColor.White,
+                        //Sampler = samplerLights,
                         Radius = 0,
-                        Intensity = 0.8f
+                        Intensity = 0.5f
                     },
                     new PointLight
                     {
                         Position = new Vector3(-1.5f, 1.5f, 1),
-                        Color = FloatColor.Yellow,
-                         Sampler = samplerLights,
+                        Color = FloatColor.White,
+                       // Sampler = samplerLights,
                         Radius = 0,
                         Intensity = 0.5f
                     },
