+++ Aethra.RayTracer/Instructions/LightsModelAndMaterials.cs
@@ -38,26 +38,20 @@ namespace Aethra.RayTracer.Instructions
 
             var reflectiveMaterial = new ReflectiveMaterial(FloatColor.White, 0.4f, 1, 300, 1f);
             var transparentMaterial = new TransparentMaterial(FloatColor.White, 0.1f, 0, 1, 3, 1);
-            var redWallMaterial = new PhongMaterial(FloatColor.Red, 1, 0, 50, 1);
-            var greenWallMaterial = new PhongMaterial(FloatColor.Green, 1, 0, 50, 1);
             var whiteWallMaterial = new PhongMaterial(FloatColor.White, 1, 0, 50, 1);
-            var emissiveMaterial = new EmissiveMaterial(FloatColor.Green, 2,
-                Texture.LoadFrom(@"_Resources/Textures/texel_density.png").ToInfo());
 
-            var reflectiveSphere = new Sphere(new Vector3(-1.25f, -1, 3), 1f, reflectiveMaterial);
-            var transparentSphere = new Sphere(new Vector3(1.25f, -1, 1), 1f, transparentMaterial);
+            var reflectiveSphere = new Sphere(new Vector3(-1.25f, 0, 3), 1f, reflectiveMaterial);
+            var transparentSphere = new Sphere(new Vector3(1.25f, 0, -1), 1f, transparentMaterial);
 
-            objects.Add(new Plane(new Vector3(-4, 0, 0), new Vector3(1, 0, 0), redWallMaterial));
-            objects.Add(new Plane(new Vector3(4, 0, 0), new Vector3(-1, 0, 0), greenWallMaterial));
+            objects.Add(new Plane(new Vector3(-4, 0, 0), new Vector3(1, 0, 0), whiteWallMaterial));
+            objects.Add(new Plane(new Vector3(4, 0, 0), new Vector3(-1, 0, 0), whiteWallMaterial));
             objects.Add(new Plane(new Vector3(5, -2, 0), new Vector3(0, 1, 0), whiteWallMaterial));
             objects.Add(new Plane(new Vector3(5, 2, 0), new Vector3(0, -1, 0), whiteWallMaterial));
             objects.Add(new Plane(new Vector3(0, 2, 6), new Vector3(0, 0, -1), whiteWallMaterial));
             objects.Add(new Plane(new Vector3(0, 2, -8), new Vector3(0, 0, 1), whiteWallMaterial));
             objects.Add(reflectiveSphere);
             objects.Add(transparentSphere);
-
-            // var crystalMaterial = new PhongMaterial(FloatColor.White, 1f, 8, 50, 1f,
-            //     Texture.LoadFrom(@"_Resources/Textures/crystal.png").ToInfo());
+            
             var crystalMaterial = new PbrMaterial(FloatColor.White, 
                 Texture.LoadFrom(@"_Resources/Textures/crystal.png").ToInfo(),
                 Texture.LoadFrom(@"_Resources/Textures/crystal.png").ToInfo(),
@@ -80,12 +74,6 @@ namespace Aethra.RayTracer.Instructions
             Scene = new Scene(objects, camera,
                 new List<Light>
                 {
-                    new PointLight
-                    {
-                        Position = new Vector3(0, 1.98f, 2),
-                        Color = FloatColor.White,
-                        // Sampler = sampler
-                    },
                     new PointLight
                     {
                         Position = new Vector3(-1, 0f, 0),
