+++ Aethra/MainWindow.xaml
@@ -49,7 +49,7 @@
             <TextBlock Grid.Row="2" Grid.Column="3" Background="{Binding Path=CreatedColor}" VerticalAlignment="Stretch" HorizontalAlignment="Stretch"></TextBlock>
             <ProgressBar x:Name="RenderProgress" Grid.Row="3" Grid.Column="0" Grid.ColumnSpan="4" VerticalAlignment="Stretch" Foreground="Green"/>
             <aethra:DrawingCanvas x:Name="Canvas" Grid.Row="4" Grid.Column="0" Grid.ColumnSpan="4" Width="800" Height="450">
-                    <aethra:CursorGraphic x:Name="CursorGraphic" Width="{Binding Path=Width, ElementName=Canvas}" Height="{Binding Path=Height, ElementName=Canvas}"/>
+                    <!-- <aethra:CursorGraphic x:Name="CursorGraphic" Width="{Binding Path=Width, ElementName=Canvas}" Height="{Binding Path=Height, ElementName=Canvas}"/> -->
             </aethra:DrawingCanvas>
 
     </Grid>
